+++ lib/vdf/MOM.cpp
@@ -252,7 +252,7 @@ float* MOM::GetDepths(){
 	for (size_t i = 0; i<_dimLens[0]*_dimLens[1]; i++){
 		if (depthsArray[i] != mv )
 			depthsArray[i] = -depthsArray[i];
-		else depthsArray[i] = MOM::vaporMissingValue();
+		else depthsArray[i] = (float)MOM::vaporMissingValue();
 	}
 	return depthsArray;
 }
@@ -781,7 +781,7 @@ void WeightTable::interp2D(const float* sourceData, float* resultData, float mis
 				else goodSum += cf2*data2;
 			if (data3 == missingValue) mvCoef += cf3;
 				else goodSum += cf3*data3;
-				if (mvCoef >= 0.5f) resultData[i+j*nlon] = MOM::vaporMissingValue();
+				if (mvCoef >= 0.5f) resultData[i+j*nlon] = (float)MOM::vaporMissingValue();
 			else resultData[i+j*nlon] = goodSum/(1.-mvCoef); 
 		}
 	}
